refactor: move plugin setup out of versioned sdk
This is one step of moving the sdk out of the cli repo.
The core uses these interfaces to create instances of plugins, regardless what language the plugins are written in.
