+++ pkg/plugin/loader/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "loader",
-    srcs = ["loader.go"],
+    srcs = [
+        "config.go",
+        "loader.go",
+    ],
     importpath = "aspect.build/cli/pkg/plugin/loader",
     visibility = ["//visibility:public"],
     deps = ["@in_gopkg_yaml_v2//:yaml_v2"],
