+++ pkg/plugin/system/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "//pkg/ioutils",
         "//pkg/plugin/client",
         "//pkg/plugin/loader",
-        "//pkg/plugin/sdk/v1alpha3/plugin",
         "//pkg/plugin/system/bep",
         "@com_github_spf13_cobra//:cobra",
         "@in_gopkg_yaml_v2//:yaml_v2",
