+++ pkg/plugin/system/system.go
@@ -27,7 +27,6 @@ import (
 	"aspect.build/cli/pkg/ioutils"
 	"aspect.build/cli/pkg/plugin/client"
 	"aspect.build/cli/pkg/plugin/loader"
-	"aspect.build/cli/pkg/plugin/sdk/v1alpha3/plugin"
 	"aspect.build/cli/pkg/plugin/system/bep"
 )
 
@@ -90,8 +89,8 @@ func (ps *pluginSystem) Configure(streams ioutils.Streams) error {
 				return fmt.Errorf("failed to configure plugin system: %w", err)
 			}
 
-			aspectPluginFile := plugin.NewAspectPluginFile(aspectpluginsPath)
-			setupConfig := plugin.NewSetupConfig(aspectPluginFile, properties)
+			aspectPluginFile := loader.NewAspectPluginFile(aspectpluginsPath)
+			setupConfig := loader.NewSetupConfig(aspectPluginFile, properties)
 			if err := aspectplugin.Setup(setupConfig); err != nil {
 				return fmt.Errorf("failed to configure plugin system: %w", err)
 			}
